move responsibility for removing permission up one level
